requirements.txt
funcy==1.17
ipdb==0.13.9
lenses==0.5.0
matplotlib==3.5.1
pandas==1.4.3
parsimonious==0.9.0
pptree==3.1
PyYAML==6.0
scikit_learn==1.1.3
scipy==1.8.1
seaborn==0.11.2
tensorboardX==2.5.1
torch_scatter==2.0.9
torchvision==0.14.0
tqdm==4.64.0
